+++ src/test/java/nextstep/cucumber/steps/auth/AuthSteps.java
@@ -1,4 +1,4 @@
-package nextstep.cucumber.steps;
+package nextstep.cucumber.steps.auth;
 
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
