+++ src/test/java/nextstep/cucumber/steps/line/SectionSteps.java
@@ -1,4 +1,4 @@
-package nextstep.cucumber.steps;
+package nextstep.cucumber.steps.line;
 
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
@@ -9,37 +9,29 @@ import org.springframework.http.MediaType;
 import java.util.HashMap;
 import java.util.Map;
 
-public class LineSteps {
-    public static ExtractableResponse<Response> 노선_생성_요청(String name, Long upstationId, Long downstationId, int distance, int duration) {
+public class SectionSteps {
+    public static ExtractableResponse<Response> 구간_추가_요청(Long lineId, Long upstationId, Long downstationId, int distance, int duration) {
         Map<String, String> params = new HashMap<>();
-        params.put("name", name);
         params.put("upstationId", upstationId.toString());
         params.put("downstationId", downstationId.toString());
-        params.put("color", "red");
         params.put("distance", Integer.toString(distance));
-        params.put("duration", duration + "");
+        params.put("duration", Integer.toString(duration));
 
         return RestAssured.given().log().all()
+                .pathParam("id", lineId)
                 .body(params)
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .when().post("/lines")
+                .when().post("/lines/{id}/sections")
                 .then().log().all()
                 .extract();
     }
 
-    public static ExtractableResponse<Response> 노선_조회(Long id) {
+    public static ExtractableResponse<Response> 구간_삭제_요청(Long lineId, Long stationId) {
         return RestAssured.given().log().all()
-                .pathParam("id", id)
+                .pathParam("id", lineId)
+                .queryParam("stationId", stationId)
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .when().get("/lines/{id}")
-                .then().log().all()
-                .extract();
-    }
-
-    public static ExtractableResponse<Response> 노선_목록_조회() {
-        return RestAssured.given().log().all()
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .when().get("/lines")
+                .when().delete("/lines/{id}/sections")
                 .then().log().all()
                 .extract();
     }
