+++ src/test/java/nextstep/cucumber/steps/station/StationSteps.java
@@ -1,4 +1,4 @@
-package nextstep.cucumber.steps;
+package nextstep.cucumber.steps.station;
 
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
