+++ src/test/resources/features/path.feature
@@ -20,7 +20,7 @@ Feature: 지하철 경로 조회 기능
       | 교대역   |
       | 남부터미널역  |
       | 양재역     |
-    And 총 거리는 5이어야 한다
+    And 최단 경로에 대한 총 거리는 5, 총 소요 시간은 5이어야 한다
 
   Scenario: 최단 경로 조회 실패
     When 사용자가 출발지와 도착지 모두를 "교대역"으로 설정하여 최단 경로를 조회하면
@@ -33,4 +33,4 @@ Feature: 지하철 경로 조회 기능
       | 교대역   |
       | 남부터미널역  |
       | 양재역     |
-    And 총 소요 시간은 5이어야 한다
+    And 최소 시간 경로에 대한 총 거리는 5, 총 소요 시간은 5이어야 한다
